+++ src/js/page/odoo/commands/common/sysparam.mjs
@@ -131,8 +131,8 @@ export default function (): Partial<CMDDef> {
     args: [
       [ARG.String, ['o', 'operation'], true, i18n.t('cmdSysParam.args.operation', "Operation to perform."), 'get', ['get', 'set', 'list']],
       [ARG.String, ['k', 'key'], false, i18n.t('cmdSysParam.args.key', 'Parameter key (required for get/set operations)')],
-      [ARG.String, ['v', 'value'], false, i18n.t('cmdSysParam.args.value', 'Parameter value (required for set operation)')],
+      [ARG.String | ARG.Number | ARG.Flag, ['v', 'value'], false, i18n.t('cmdSysParam.args.value', 'Parameter value (required for set operation)')],
     ],
-    example: `-o get -k account_online_synchronization.proxy_mode\n-o set -k account_online_synchronization.proxy_mode -v sandbox\n-o list`,
+    example: '-o set -k account_online_synchronization.proxy_mode -v sandbox',
   };
 }