+++ src/js/page/terminal/utils/pretty_object_string.mjs
@@ -3,7 +3,8 @@
 // License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 import encodeHTML from './encode_html';
+import replacer from './stringify_replacer';
 
 export default function (obj: {...}): string {
-  return encodeHTML(JSON.stringify(obj, null, 4));
+  return encodeHTML(JSON.stringify(obj, replacer, 4));
 }